shared/units: Move unit functions
